requirements.txt
langchain
langchain-community
langchain-openai
chromadb
python-dotenv
streamlit
groq
langchain-huggingface